Refactor imports: update model imports to type imports for better clarity and organization
