+++ src/types/tech.ts
